fix, optimizations and new client packets

+++ core/src/ISidedBlock.java
@@ -3,5 +3,5 @@ package com.qzx.au.core;
 import net.minecraft.world.World;
 
 public interface ISidedBlock {
-	public abstract void toggleSide(World world, int x, int y, int z, int side);
+	public abstract void toggleSide(World world, int x, int y, int z, byte side);
 }
